Bug 994976 - separate test logic calendar with and without mocks
