+++ apps/calendar/test/marionette/alarm_test.js
@@ -9,84 +9,48 @@ marionette('alarm', function() {
   var app;
   var client = marionette.client();
 
-  suite(' > with alarm mocks', function() {
+  setup(function() {
+    app = new Calendar(client);
+    client.contentScript.inject(SHARED_PATH + '/mock_navigator_mozalarms.js');
+    client.contentScript.inject(SHARED_PATH +
+                                '/mock_navigator_moz_set_message_handler.js');
+    app.launch({ hideSwipeHint: true });
+  });
 
+  suite('create event with a single reminder', function() {
     setup(function() {
-      app = new Calendar(client);
-      client.contentScript.inject(SHARED_PATH +
-        '/mock_navigator_mozalarms.js');
-      client.contentScript.inject(SHARED_PATH +
-        '/mock_navigator_moz_set_message_handler.js');
-      app.launch({ hideSwipeHint: true });
+      createEventWithReminders(app, ['5 minutes before']);
+      app.monthDay.scrollToEvent();
     });
 
-    suite('create event with a single reminder', function() {
-      setup(function() {
-        createEventWithReminders(app, ['5 minutes before']);
-        app.monthDay.scrollToEvent();
-      });
-
-      test('should display reminder in read-only event view', function() {
-        var reminders = getEventReminders(app);
-        assert.equal(reminders.length, 1);
-        assert.include(reminders, '5 minutes before');
-      });
-
-      test('should add alarm to alarms api', function() {
-        var alarms = getMozAlarms(client);
-        assert.equal(alarms.length, 1);
-      });
+    test('should display reminder in read-only event view', function() {
+      var reminders = getEventReminders(app);
+      assert.equal(reminders.length, 1);
+      assert.include(reminders, '5 minutes before');
     });
 
-    suite('create an event with two valid reminders', function() {
-      setup(function() {
-        createEventWithReminders(app, ['5 minutes before',
-                                       '15 minutes before']);
-        app.monthDay.scrollToEvent();
-      });
-
-      test('should display both reminders in read-only view', function() {
-        var reminders = getEventReminders(app);
-        assert.equal(reminders.length, 2);
-        assert.include(reminders, '5 minutes before');
-        assert.include(reminders, '15 minutes before');
-      });
-
-      test('should only fire one alarm', function() {
-        var alarms = getMozAlarms(client);
-        assert.equal(alarms.length, 1);
-      });
+    test('should add alarm to alarms api', function() {
+      var alarms = getMozAlarms(client);
+      assert.equal(alarms.length, 1);
     });
   });
 
-  suite('> without alarm mocks', function() {
+  suite('create an event with two valid reminders', function() {
     setup(function() {
-      app = new Calendar(client);
-      app.launch({ hideSwipeHint: true });
+      createEventWithReminders(app, ['5 minutes before', '15 minutes before']);
+      app.monthDay.scrollToEvent();
     });
 
-    suite('create two events with simultaneous reminders', function() {
-      setup(function() {
-        var startTime = Date.now() + 5000; // 5 seconds in future
-        app.createEvent({
-          title: 'First Event',
-          startDate: new Date(startTime),
-          reminders: ['At time of event']
-        });
-        app.createEvent({
-          title: 'Second Event',
-          startDate: new Date(startTime),
-          reminders: ['At time of event']
-        });
-      });
+    test('should display both reminders in read-only view', function() {
+      var reminders = getEventReminders(app);
+      assert.equal(reminders.length, 2);
+      assert.include(reminders, '5 minutes before');
+      assert.include(reminders, '15 minutes before');
+    });
 
-      test('both events trigger notifications', function() {
-        client.switchToFrame();
-        client.waitFor(function() {
-          var notificationSelector = '#desktop-notifications-container > div';
-          return client.findElements(notificationSelector).length > 1;
-        });
-      });
+    test('should only fire one alarm', function() {
+      var alarms = getMozAlarms(client);
+      assert.equal(alarms.length, 1);
     });
   });
 });
