+++ shared/test/integration/tbpl-manifest.json
@@ -1,6 +1,5 @@
 {
   "blacklist": [
-    "apps/calendar/test/marionette/alarm_test.js",
     "apps/calendar/test/marionette/caldav_test.js",
     "apps/clock/test/marionette/alarm_test.js",
     "apps/communications/contacts/test/marionette/details_test.js",
