+++ shared/test/integration/travis-manifest.json
@@ -1,6 +1,5 @@
 {
   "blacklist": [
-    "apps/calendar/test/marionette/alarm_test.js",
     "apps/calendar/test/marionette/caldav_test.js",
     "apps/clock/test/marionette/alarm_test.js",
     "apps/email/test/marionette/activity_create_email_account_complete_test.js",
